Fix Grouping Again
The modified logic was selecting lines that started with the project
root. However, the line object looks like this

{
  "number"=>"148",
  "file"=>"[PROJECT_ROOT]/app/views/shared/_grid_block_item.html.erb",
  "method"=> "block in _app_views_shared__grid_block_item_html_erb___4054598342690822596_69862626568500"
}

The method part sometimes is dynamic / changes. When this is hashed is
causes the same "error" to spit out a different hash.

To fix just using the "file" part when computing the hash.